gcp/requirements.txt
tensorflow
google-cloud-storage
Pillow
flask==1.1.4
flask_cors==3.0.10
markupsafe==1.1.1

